Refactoring: common auth
